requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
reportlab==4.0.4
python-dotenv==1.0.0
Flask-Login==0.6.3
Jinja2==3.1.2
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-WTF==1.1.1
PyMySQL==1.1.0
cryptography==41.0.7